+++ analysis_driver/quality_control/__init__.py
@@ -5,3 +5,4 @@ from .median_coverage import SamtoolsDepth
 from .contamination_blast import ContaminationBlast
 from .calculate_relatedness import Relatedness, Peddy, Genotype_gVCFs
 from .bcl_validation import BCLValidator
+from .detect_bad_cycles_tiles import BadTileCycleDetector