+++ CHANGELOG.md
@@ -1,16 +1,35 @@
 Changelog for Analysis-Driver
 =============================
 
-0.15 (unreleased)
+0.16 (unreleased)
 -----------------
 
 - Nothing changed yet.
 
 
+0.15.2 (2017-05-29)
+-------------------
+
+- Change tools writting to /tmp in Samtools depth and Genotype Validation.
+
+
+0.15.1 (2017-05-24)
+-------------------
+
+- Temporary fix to make sure the fastq_filterer stats file is present to be parsed by RunCrawler.
+
+
+0.15 (2017-05-18)
+-----------------
+
+ - All sample and project processes are now segmented using Luigi
+ - Allow filtering/Trimming of bad quality runs in demultiplexing 
+ - Fix analysis driver --stop and Error handling
+
 0.14.3 (2017-04-28)
 -------------------
 
-- Removed need for SampleSheet to exist for a Run to be picked up
+ - Removed need for SampleSheet to exist for a Run to be picked up
 
 
 0.14.2 (2017-04-21)